+++ resources/js/i18n/locales/vi/reports.js
@@ -484,7 +484,7 @@ export default {
         product: 'Sản phẩm',
         warehouse: 'Kho',
         contacts: 'Số contact',
-        raw_packets: 'Gói tin raw',
+        raw_packets: 'Gói tin nhận được',
         primary_packets: 'Gói tin chính',
         upsale_packets: 'Gói tin upsale',
         contacts_total: 'Tổng contact',